+++ lib/src/charts/data/bars/multi_bar.dart
@@ -23,10 +23,9 @@ class MultiBar extends BarGroup {
 
   /// The space between consecutive bars in a series arrangement.
   ///
-  /// **Note:** If you've chosen [arrangement] as stack, then do not use
-  /// groupSpacing. Providing a value higher than zero in the case
-  /// will throw an Exception.
-  final double groupSpacing;
+  /// **Note:** If you've chosen [arrangement] as stack, then [spacing] will
+  /// be ignored.
+  final double spacing;
 
   /// Defines a Group of Multiple Bars
   ///
@@ -35,41 +34,35 @@ class MultiBar extends BarGroup {
   ///
   /// See Also: [BarGroup]
   ///
-  MultiBar({
-    required super.xValue,
-    required this.yValues,
-    // defaults to series i.e. side by side
-    this.arrangement = BarGroupArrangement.series,
-    this.groupSpacing = 0.0,
-    super.groupStyle,
-  })  : assert(yValues.isNotEmpty, "At least one yValue is required!"),
-        // Ensure that groupSpacing is not applied when arrangement is stack
-        assert(
-          (arrangement == BarGroupArrangement.stack && groupSpacing > 0.0)
-              ? false
-              : true,
-          "groupSpacing should not be used when Bar Arrangement is Stack!",
-        ),
-        assert(groupSpacing >= 0.0, "groupSpacing cannot be Negative!");
+  MultiBar(
+      {required super.xValue,
+      required this.yValues,
+      // defaults to series i.e. side by side
+      this.arrangement = BarGroupArrangement.series,
+      this.spacing = 10.0,
+      super.style,
+      super.padding})
+      : assert(yValues.isNotEmpty, "At least one yValue is required!"),
+        assert(spacing >= 0.0, "groupSpacing cannot be Negative!");
 
   /// Lerps between two [MultiBar]s for a factor [t]
   static MultiBar lerp(BarGroup? current, BarGroup target, double t) {
     if ((current is MultiBar?) && target is MultiBar) {
       return MultiBar(
-        xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
-        yValues: BarData.lerpBarDataList(current?.yValues, target.yValues, t),
-        arrangement: target.arrangement,
-        groupSpacing: lerpDouble(
-          current?.groupSpacing,
-          target.groupSpacing,
-          t,
-        ).asOrDefault(0.0),
-        groupStyle: BarDataStyle.lerp(
-          current?.groupStyle,
-          target.groupStyle,
-          t,
-        ),
-      );
+          xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
+          yValues: BarData.lerpBarDataList(current?.yValues, target.yValues, t),
+          arrangement: target.arrangement,
+          spacing: lerpDouble(
+            current?.spacing,
+            target.spacing,
+            t,
+          ).asOrDefault(0.0),
+          style: BarDataStyle.lerp(
+            current?.style,
+            target.style,
+            t,
+          ),
+          padding: lerpDouble(current?.padding, target.padding, t) ?? 0.0);
     } else {
       throw Exception('Both current & target data should be of same series!');
     }
@@ -77,5 +70,5 @@ class MultiBar extends BarGroup {
 
   @override
   List<Object?> get props =>
-      [xValue, yValues, groupStyle, groupSpacing, arrangement];
+      [xValue, yValues, style, spacing, arrangement, padding];
 }